+++ src/components/collections/Header/utils/getLogoMarkup.js
@@ -4,14 +4,21 @@ import { Menu, Image } from 'semantic-ui-react';
 import { Heading } from 'typography/Heading';
 
 /**
+ * @param  {string} logoHref
  * @param  {string} logoText
  * @param  {string} logoSrc
  * @param  {string} logoSizes
  * @param  {string} logoSrcSet
  * @return {Object}
  */
-export const getLogoMarkup = (logoText, logoSrc, logoSizes, logoSrcSet) => (
-  <Menu.Item href="/" link>
+export const getLogoMarkup = (
+  logoHref,
+  logoText,
+  logoSrc,
+  logoSizes,
+  logoSrcSet
+) => (
+  <Menu.Item href={logoHref} link>
     {logoSrc ? (
       <Image
         alt={logoText}