+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [1.147.0](https://github.com/lodgify/lodgify-ui/compare/v1.146.0...v1.147.0) (2019-03-01)
+
+
+### Features
+
+* **Header:** expose logoHref prop ([bc70f92](https://github.com/lodgify/lodgify-ui/commit/bc70f92))
+* **Hero:** pass through headerLogoHref to Header ([742ef3a](https://github.com/lodgify/lodgify-ui/commit/742ef3a))
+* **HomepageHero:** pass through headerLogoHref to Hero ([c6e3181](https://github.com/lodgify/lodgify-ui/commit/c6e3181))
+* **PropertyPageHero:** pass through headerLogoHref to Hero ([914bdbd](https://github.com/lodgify/lodgify-ui/commit/914bdbd))
+
 # [1.146.0](https://github.com/lodgify/lodgify-ui/compare/v1.145.0...v1.146.0) (2019-03-01)